requirements.txt
colorlog==6.7.0
homeassistant==2023.7.3
pip>=21.0,<23.3
ruff==0.0.279
paho-mqtt~=1.6.1
voluptuous~=0.13.1

dacite~=1.8.1